BugFix: Can't join arena agian after finishing
People could not join arenas again after finishing? @FaberoM added a playerArenaMap which I don't get why?? It never gets updated when a player leaves an arena and tehrefore create a bug. This method also seems reduntant as all the information needed is already in the getPlayers() from the arena object.

This is a temporary fix. I think it's best to remove playerArenaMap entirely.