Fix async catch exception in old versions when teleporting players to their plots
Also there is chances when some users are tests bb in flat map without any building, and after the player attempting to teleport to the plot, he never get to that plot. Now on newest versions the cuboid calculations will doing in async thread, in old versions we stay on main thread for successful teleport.